src/pyproject.toml
[tool.poetry]
name = "src"
version = "1.20"
description = ""
authors = ["kyoF <fujiki.kyosuke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^2.2.3"
Flask-SQLAlchemy = "^3.0.3"
Flask-Migrate = "^4.0.4"
pymysql = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
